# server/openapi.yaml
openapi: 3.0.1
info:
  title: MMOCG
  description: 'This is the Massive Multiplayer Online Clicker Game server behind 
  [emoji-clicker](https://github.com/fabjan/emoji-clicker).'
  contact:
    name: Fabian Bergström
    url: https://github.com/fabjan
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
  version: 1.0.0
externalDocs:
  description: Find out more about Emoji Clicker
  url: https://github.com/fabjan/emoji-clicker
servers:
- url: https://example.com/v1
tags:
- name: team
  description: Registering, inspecting teams
- name: clicks
  description: Competing
paths:
  /leaderboard:
    get:
      tags:
      - team
      summary: Returns the highest scoring teams
      operationId: getLeaderboard
      responses:
        200:
          description: Leaderboard found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
  /team/{teamId}:
    post:
      tags:
      - team
      summary: Creates or updates a team with form data
      operationId: updateTeam
      parameters:
      - name: teamId
        in: path
        description: ID of team to update
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid team ID
        200:
          description: Team exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        201:
          description: Team was created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
    get:
      tags:
      - team
      summary: Find team by ID
      description: Returns a single team
      operationId: getTeamById
      parameters:
      - name: teamId
        in: path
        description: ID of team to find
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid team ID
        404:
          description: Team not found
        200:
          description: Team found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
  /team/{teamId}/click:
    post:
      tags:
      - clicks
      summary: Report clicks for the given team 
      operationId: click
      parameters:
      - name: teamId
        in: path
        description: ID of team to update
        required: true
        schema:
          type: string
      - name: count
        in: query
        description: Amount of clicks to report
        schema:
          type: integer
          minimum: 1
          maximum: 10
      responses:
        400:
          description: Invalid team ID
        402:
          description: Invalid click count
        404:
          description: Team not found
        429:
          description: Stop cheating
        200:
          description: Click(s) were recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
components:
  schemas:
    Leaderboard:
      type: array
      items:
        $ref: '#/components/schemas/Team'
    Team:
      type: object
      properties:
        id:
          type: string
        clicks:
          type: integer
          format: int64
          minimum: 0
          maximum: 9007199254740992